+++ snapboost/__init__.py
@@ -1,5 +1,6 @@
 from hnbm import HNBM
 
+from .kernel_learner import WeightedKernelRidgeRegressor
 from .linear_learner import WeightedLinearRegressor
 from .preprocessing import make_tabular_preprocessor
 from .rff_learner import LaplacianSampler, RandomFourierRidgeRegressor
@@ -16,6 +17,7 @@ __all__ = [
     "HNBM",
     "LaplacianSampler",
     "make_tabular_preprocessor",
+    "WeightedKernelRidgeRegressor",
     "WeightedLinearRegressor",
     "RandomFourierRidgeRegressor",
     "SnapBoost",
@@ -26,4 +28,4 @@ __all__ = [
     "SnapBoost_KernelRidge",
 ]
 
-__version__ = "1.0.0"
+__version__ = "1.2.0"